setup basic build infrastructure
This is an evolution over the setup used by Limes, swift-http-import
etc. It fully embraces Go modules and uses an *empty* GOPATH instead of
a synthetic one. GOPATH is still set to $PWD/.gopath/, but only to
ensure a deterministic placement of the build cache (at
$PWD/.gopath/pkg). If the GOPATH variable were empty, Go would put that
stuff at $HOME/go/pkg which is urgh.