rewrite GetMatchingReasons into a constraint-solver thingy
This appears to be the best (i.e., least worst) design where
percentage-step resizing and single-step resizing can happily coexist
with the various constraints that we already support. And it feels
extensible enough to support new usecases that come later.

I expect that this breaks some asset-scrape tests, not only because I
haven't checked for bugs, but because the percentage-step resizing now
gets a bit more wiggle room: For example, when we want to upsize by 10%,
but hit the maximum size after 5%, we used to not resize at all. Now the
10% increase is only the "desired" value and the constraint solver will
choose a smaller value if all the constraints are still satisfied. This
is something where we need extra test coverage.

However, it's weekend already, so I will tackle the tests next week.