constrain resize operations to not jump between threshold
It used to be possible, with an excessive step size, to jump from above
the high threshold straight below the low threshold, causing Castellum
to generate an endless loop of upsizing and downsizing operations. To
avoid this, we now register the opposite threshold(s) as a constraint
for assets reporting absolute usage.