add optional additional logging for scrape results
We're investigating a case where Castellum mysteriously decided not to
create a "critical" operation on an NFS share even though the NetApp
metrics say that it should have done so. There are no error logs
anywhere along the way. With this additional logging, we should be able
to better understand why Castellum chose to do nothing here.