rewrite tasks.ScrapeNextAsset to not take `assetType` as a parameter
I wanted to have this for some time anyway in order to better distribute
the load of asset scraping. At this point, this change is unavoidable as
a prerequisite for an upcoming new asset manager, which will not have a
fixed set of asset types. Since we will not be able to enumerate asset
types at startup, the previous design was not tenable anymore.

I will also make the same change to ScrapeNextResource soon.