add CASTELLUM_INSECURE flag, refactor ProjectScopedClient
In the server-groups plugin, we need several ServiceClients at once in
the same project scope, so it makes more sense to return a
ProviderClient from func ProjectScopedClient.

Also, there was a bug in ProjectScopedClient where it always failed on
fresh projects because zero role assignments always results in 401. This
is also fixed in this commit. All of this happened on the same feature
branch and I'm now too lazy to untangle these three separate things into
their own commits.