add Asset.MinimumSize and Asset.MaximumSize
We will need this for the nfs-shares asset type. The actual NetApp
shares corresponding to Manila shares include the share's snapshots with
its size (in order to leverage deduplication), so even though the usage
should match what the user sees in `df` and such, we should not resize
below the total usage that is comprised out of the actual share usage
and the additional snapshot usage.