core/logic: clarify the distinction between limiting and enforceable constraints
The existing MinimumSize/MaximumSize constraints on resources are
designed and documented as limiting constraints that do not cause
action, e.g. the UI shows them as "Never extend shares above [___] GiB"
and such.

By contrast, the existing MinimumFreeSize constraint on resources is
designed (but was not yet documented) as a constraint that can cause
action, e.g. the UI shows it as "Ensure at least [___] GiB free".

The relatively new asset-level MinimumSize/MaximumSize constraints shall
follow in this second category, which I now start calling enforceable
constraints. To better distinguish these from the purely limiting
constraints on the resource configuration, these constraints are renamed
to StrictMinimumSize and StrictMaximumSize in the code.

This is the better version of the series of commits from last week where
I set out to implement this enforcing behavior, but did not take into
account that the existing constraints are documented as merely limiting,
and that users are relying on this semantic.